+++ examples/cuda/vector_add_tegra_observer.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python
 """This is the minimal example from the README"""
 
+import json
+
 import numpy
 from kernel_tuner import tune_kernel
-from kernel_tuner.file_utils import store_output_file, store_metadata_file
-import logging
-from collections import OrderedDict
+from kernel_tuner.observers.tegra import TegraObserver
 
 def tune():
 
@@ -18,7 +18,7 @@ def tune():
     }
     """
 
-    size = 10000000
+    size = 800000
 
     a = numpy.random.randn(size).astype(numpy.float32)
     b = numpy.random.randn(size).astype(numpy.float32)
@@ -27,14 +27,17 @@ def tune():
 
     args = [c, a, b, n]
 
-    tune_params = OrderedDict()
+    tune_params = dict()
     tune_params["block_size_x"] = [128+64*i for i in range(15)]
 
-    results, env = tune_kernel("vector_add", kernel_string, size, args, tune_params, lang="HIP",
-                               log=logging.DEBUG)
+    tegraobserver = TegraObserver(["core_freq"])
 
-    # Store the metadata of this run
-    store_metadata_file("vector_add-metadata.json")
+    metrics = dict()
+    metrics["f"] = lambda p: p["core_freq"]
+
+    results, env = tune_kernel("vector_add", kernel_string, size, args, tune_params, observers=[tegraobserver], metrics=metrics)
+
+    print(results)
 
     return results